working ish using reset field pos

+++ src/main/java/frc/robot/commands/PlaceCone/AllLime.java
@@ -20,7 +20,7 @@ public class AllLime extends SequentialCommandGroup {
       DrivetrainSubsystem drivetrainSubsystem,
       PoseEstimatorSubsystem poseEstimatorSubsystem,
       LimeLightSub limeLightSub,
-      int degrees) {
+      double degrees) {
     // Add your commands in the addCommands() call, e.g.
     // addCommands(new FooCommand(), new BarCommand());
     addCommands(
